modified check for short files.
The check now sends any file with tones detected less than 30 seconds to a queue, the next audio file from that talkgroup gets appended to it and it will use the audio metadata from the first post.

This is to avoid Dispatchers that take a long time to send a transmission.